Move code outside of ColorSpaceConverter into their own files
